Generate confxml sources via the build system
Removes the confxml.{c,h} from git and adds targets to automatically
generate them at build time. Removes the error-prone manual step of
having to generate them manually and commit to git.

The binary data commited to git are hard to verify for correctness and
is in essence a duplication of information.