Add Package for the Dive Xtras Blacktip / CudaX DPV.
Add a package that contains an improved version of the runtime and UI
for the Dive Xtras Blacktip / CudaX DPVs (a.k.a. 'dive scooter').

https://dive-xtras.com/pages/blacktip-series

This is an improved version of the runtime provided by the manufacturer
that has been updated to run on VESC 6.06.
For more information see `blacktip_dpv/README.md`.

Signed-off-by: Michael Keller <github@ike.ch>